Adds createTask feature
Users can now create a task which is assigned to a Project by the
project_id

Also, uses jQuery to populate the project_id input of the createTask
form upon createProjectSuccess.  This will allow the user to create a
task without needing to know the ID of the project.

Next step is to add feature to allow users to create multiple tasks.