Adds deleteTask function
Users can delete tasks now, once they're complete.  This is not the
final iteration, as I want users to be able to review the tasks,
but I need a more eloquent way to find the "iscomplete" status.

Next step is to clean up page so it's a better user experience.