Tidies up html and some bugs within code
Not sure what happened, will need to read more on Git, but I was getting
phantom merge conflicts with random text like "<<<<<<<<< HEAD" and the
names of my branches showing up in the DOM, but it was not in any of my
files.

The fix for this was just to go into the file that my grunt server was
telling me had errors in it, find a random empty line, hit 'space', then
'delete', then saved the file.  All the text went away, and the pages
loaded without error.

Not sure what happened, but will research this.